refactor: render관련 책임 컨트롤러에서 webOutputView로 이동
