+++ src/main/java/chess/controller/WebChessController.java
@@ -9,24 +9,16 @@ import static spark.Spark.stop;
 
 import chess.dao.BoardDao;
 import chess.dao.GameDao;
-import chess.domain.ChessBoardInitLogic;
 import chess.domain.ChessBoardPosition;
 import chess.domain.ChessGame;
 import chess.domain.Team;
 import chess.domain.piece.ChessPiece;
 import chess.service.StorageService;
-import chess.webview.ChessPieceImagePath;
-import chess.webview.ColumnName;
-import chess.webview.RowName;
 import chess.webview.WebInputView;
 import chess.webview.WebOutputView;
-import java.util.HashMap;
 import java.util.Map;
-import java.util.Map.Entry;
-import spark.ModelAndView;
 import spark.Request;
 import spark.Response;
-import spark.template.handlebars.HandlebarsTemplateEngine;
 
 public class WebChessController {
     private static final int GAME_ID = 1111;
@@ -49,7 +41,7 @@ public class WebChessController {
         });
 
         get("/board", (req, res) -> {
-            return render(makeBoardModel(storageService.loadChessBoardData(chessGame.getGameId())), "board.html");
+            return WebOutputView.goBoardPage(storageService.loadChessBoardData(chessGame.getGameId()));
         });
 
         post("/inGameCommand", (req, res) -> {
@@ -63,8 +55,7 @@ public class WebChessController {
         });
 
         get("/status", (req, res) -> {
-            Map<String, Object> model = WebOutputView.makeStatusModel(chessGame.getTeamScore(), chessGame.getWinner());
-            return render(model, "status.html");
+            return WebOutputView.goStatusPage(chessGame.getTeamScore(), chessGame.getWinner());
         });
 
         exception(Exception.class, (exception, request, response) -> {
@@ -80,14 +71,6 @@ public class WebChessController {
         chessGame.initialize(turn, mapData);
     }
 
-    public static Map<String, Object> makeBoardModel(Map<ChessBoardPosition, ChessPiece> mapData) {
-        Map<String, Object> boardModel = new HashMap<>();
-        for (Entry<ChessBoardPosition, ChessPiece> entry : mapData.entrySet()) {
-            boardModel.put(chessBoardToString(entry.getKey()), ChessPieceImagePath.of(entry.getValue()));
-        }
-        return boardModel;
-    }
-
     private static void doMoveCommand(Request req, ChessGame chessGame, StorageService storageService) {
         ChessBoardPosition source = WebInputView.extractSource(req.queryParams("command"));
         ChessBoardPosition target = WebInputView.extractTarget(req.queryParams("command"));
@@ -97,17 +80,9 @@ public class WebChessController {
 
     private static String goFirstPageIfGameEnd(Response res, ChessGame chessGame) {
         if (chessGame.isGameEnd()) {
-            return render(null, "../public/index.html");
+            return WebOutputView.goInitialPage();
         }
         res.redirect("/board");
         return null;
-    }
-
-    private static String chessBoardToString(ChessBoardPosition chessBoardPosition) {
-        return ColumnName.of(chessBoardPosition.getColumn()) + RowName.of(chessBoardPosition.getRow());
-    }
-
-    private static String render(Map<String, Object> model, String templatePath) {
-        return new HandlebarsTemplateEngine().render(new ModelAndView(model, templatePath));
     }
 }
